requirements.txt
hydra-core==1.2.0
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
pytorch-lightning==1.6.3
seaborn==0.11.2
torch==1.11.0
torchvision==0.12.0
wandb==0.12.17
loguru==0.6.0
